+++ src/chain/tezos/contracts/BabylonDelegationHelper.ts
@@ -1,4 +1,5 @@
 import * as blakejs from 'blakejs';
+import { JSONPath } from 'jsonpath-plus';
 
 import { TezosLanguageUtil } from '../TezosLanguageUtil';
 import { TezosNodeReader } from '../TezosNodeReader';
@@ -34,9 +35,15 @@ export namespace BabylonDelegationHelper {
     export function verifyScript(script: string): boolean {
         const k = Buffer.from(blakejs.blake2s(TezosLanguageUtil.preProcessMichelsonScript(script).join('\n'), null, 16)).toString('hex');
 
-        if (k !== '023fc21b332d338212185c817801f288') { throw new Error(`Contract does not match the expected code hash: ${k}, '023fc21b332d338212185c817801f288'`); }
+        if (k !== 'a585489ffaee60d07077059539d5bfc8') { throw new Error(`Contract does not match the expected code hash: ${k}, 'a585489ffaee60d07077059539d5bfc8'`); }
 
         return true;
+    }
+
+    export async function getSimpleStorage(server: string, address: string): Promise<{administrator: string}> {
+        const storageResult = await TezosNodeReader.getContractStorage(server, address);
+
+        return { administrator: JSONPath({ path: '$.string', json: storageResult })[0] };
     }
 
     /**